Видимо не заметил в прошлый раз транзакцию. Внутри нее лучше вводом-выводом не заниматься. Ну, по крайней мере, ожидать ввода пользователя
