+++ spring-05/testing-students/src/main/java/ru/otus/testing/commandline/BookCommandLine.java
@@ -3,21 +3,38 @@ package ru.otus.testing.commandline;
 import org.springframework.shell.standard.ShellComponent;
 import org.springframework.shell.standard.ShellMethod;
 import ru.otus.testing.service.BookService;
+import ru.otus.testing.service.IOService;
+import ru.otus.testing.service.UserAnswerService;
 
 @ShellComponent
 public class BookCommandLine {
 
     private final BookService bookService;
 
+    private final IOService ioService;
 
-    public BookCommandLine(BookService bookService) {
+    private final UserAnswerService userAnswerService;
+
+    public BookCommandLine(BookService bookService, IOService ioService, UserAnswerService userAnswerService) {
         this.bookService = bookService;
+        this.ioService = ioService;
+        this.userAnswerService = userAnswerService;
     }
 
 
     @ShellMethod(value = "create-book", key = {"create-book", "-c-book"})
     public String create() {
-        bookService.create();
+        ioService.outputString("Enter books info, please:");
+
+        var bookName = ioService.readNextWithPrompt("- Enter book name: ");
+        var bookYear = userAnswerService.checkUserAnswer("- Enter book written year: ");
+
+        var authorName = ioService.readNextWithPrompt("- Enter author name: ");
+        var authorYear = userAnswerService.checkUserAnswer("- Enter author years: ");
+
+        var genreName = ioService.readNextWithPrompt("- Enter books genre name: ");
+
+        bookService.create(bookName, bookYear, authorName, authorYear, genreName);
         return "Book was created";
     }
 
@@ -35,7 +52,19 @@ public class BookCommandLine {
 
     @ShellMethod(value = "update-book", key = {"update-book", "-u-book"})
     public String update() {
-        bookService.update();
+        ioService.outputString("Enter books info, please: ");
+
+        var bookId = userAnswerService.checkUserAnswer("- Enter book id, that you want update: ");
+
+        var bookName = ioService.readNextWithPrompt("- Enter book name: ");
+        var bookYear = userAnswerService.checkUserAnswer("- Enter book written year: ");
+
+        var authorName = ioService.readNextWithPrompt("- Enter author name: ");
+        var authorYear = userAnswerService.checkUserAnswer("- Enter author years: ");
+
+        var genreName = ioService.readNextWithPrompt("- Enter books genre name: ");
+
+        bookService.update(bookId, bookName, bookYear, authorName, authorYear, genreName);
         return "Info about book was updated";
     }
