+++ spring-05/testing-students/src/main/java/ru/otus/testing/service/BookService.java
@@ -6,13 +6,13 @@ import ru.otus.testing.model.Book;
 import java.util.List;
 
 public interface BookService {
-    Book create();
+    Book create(String bookName, long bookYear, String authorName, long authorYear, String genreName);
 
     Book readById();
 
     List<Book> readAll();
 
-    void update();
+    void update(long bookId, String bookName, long bookYear, String authorName, long authorYear, String genreName);
 
     void delete();
 }
